detect-fake-imagee/requirements.txt
transformers
torch
torchvision
pillow
sentencepiece
protobuf
